+++ package.json
@@ -65,8 +65,7 @@
     "icon_url": "logo.jpg",
     "disabled": false,
     "dependencies": {
-        "@azure/msal-node": "1.18.4",
-        "axios": "1.12.0",
+        "@hackolade/fetch": "1.3.0",
         "base64url": "3.0.1",
         "crypto": "1.0.1",
         "lodash": "4.17.21",
