+++ reverse_engineering/databaseService/databaseService.js
@@ -1,147 +1,10 @@
 const { flatMap } = require('lodash');
-const axios = require('axios');
 const fs = require('fs');
 const https = require('https');
-const msal = require('@azure/msal-node');
 const sql = require('mssql');
 const { getObjectsFromDatabase, getNewConnectionClientByDb } = require('./helpers');
 const getSampleDocSize = require('../helpers/getSampleDocSize');
-
-const QUERY_REQUEST_TIMEOUT = 60000;
-
-const getSslConfig = ({ connectionInfo }) => {
-	if (connectionInfo.sslType === 'SYSTEMCA') {
-		return {};
-	}
-
-	if (connectionInfo.sslType === 'TRUST_ALL_CERTIFICATES') {
-		return {
-			trustServerCertificate: true,
-		};
-	}
-
-	if (connectionInfo.sslType === 'TRUST_CUSTOM_CA_SIGNED_CERTIFICATES') {
-		return {
-			cryptoCredentialsDetails: {
-				ca: fs.readFileSync(connectionInfo.certAuthority),
-			},
-		};
-	}
-
-	if (connectionInfo.sslType === 'TRUST_SERVER_CLIENT_CERTIFICATES') {
-		return {
-			cryptoCredentialsDetails: {
-				ca: fs.readFileSync(connectionInfo.certAuthority),
-				cert: connectionInfo.clientCert && fs.readFileSync(connectionInfo.clientCert),
-				key: connectionInfo.clientPrivateKey && fs.readFileSync(connectionInfo.clientPrivateKey),
-				passphrase: connectionInfo.passphrase,
-			},
-		};
-	}
-
-	return {};
-};
-
-const getConnectionClient = async ({ connectionInfo, logger }) => {
-	const hostName = getHostName(connectionInfo.host);
-	const userName =
-		isEmail(connectionInfo.userName) && hostName
-			? `${connectionInfo.userName}@${hostName}`
-			: connectionInfo.userName;
-	const tenantId = connectionInfo.connectionTenantId || connectionInfo.tenantId || 'common';
-	const sslOptions = getSslConfig({ connectionInfo });
-	const clientId = '0dc36597-bc44-49f8-a4a7-ae5401959b85';
-	const redirectUri = 'http://localhost:8080';
-	const timeout = Number(connectionInfo.queryRequestTimeout) || 60000;
-
-	if (connectionInfo.authMethod === 'Username / Password') {
-		return await sql.connect({
-			user: userName,
-			password: connectionInfo.userPassword,
-			server: connectionInfo.host,
-			port: +connectionInfo.port,
-			database: connectionInfo.databaseName,
-			options: {
-				enableArithAbort: true,
-				encrypt:
-					connectionInfo.encryptConnection === undefined ? true : Boolean(connectionInfo.encryptConnection),
-				...sslOptions,
-			},
-			connectTimeout: timeout,
-			requestTimeout: timeout,
-		});
-	} else if (connectionInfo.authMethod === 'Username / Password (Windows)') {
-		return await sql.connect({
-			user: userName,
-			password: connectionInfo.userPassword,
-			server: connectionInfo.host,
-			port: +connectionInfo.port,
-			database: connectionInfo.databaseName,
-			domain: connectionInfo.userDomain,
-			options: {
-				...sslOptions,
-				encrypt:
-					connectionInfo.encryptWindowsConnection === undefined
-						? false
-						: Boolean(connectionInfo.encryptWindowsConnection),
-				enableArithAbort: true,
-			},
-			connectTimeout: timeout,
-			requestTimeout: timeout,
-		});
-	} else if (connectionInfo.authMethod === 'Azure Active Directory (MFA)') {
-		const token = await getToken({ connectionInfo, tenantId, clientId, redirectUri, logger });
-
-		return await sql.connect({
-			server: connectionInfo.host,
-			port: +connectionInfo.port,
-			database: connectionInfo.databaseName,
-			options: {
-				...sslOptions,
-				encrypt: true,
-				enableArithAbort: true,
-			},
-			authentication: {
-				type: 'azure-active-directory-access-token',
-				options: {
-					token,
-				},
-			},
-			connectTimeout: QUERY_REQUEST_TIMEOUT,
-			requestTimeout: QUERY_REQUEST_TIMEOUT,
-		});
-	} else if (connectionInfo.authMethod === 'Azure Active Directory (Username / Password)') {
-		return await sql.connect({
-			user: userName,
-			password: connectionInfo.userPassword,
-			server: connectionInfo.host,
-			port: +connectionInfo.port,
-			database: connectionInfo.databaseName,
-			options: {
-				...sslOptions,
-				encrypt: true,
-				enableArithAbort: true,
-			},
-			authentication: {
-				type: 'azure-active-directory-password',
-				options: {
-					userName: connectionInfo.userName,
-					password: connectionInfo.userPassword,
-					tenantId,
-					clientId,
-				},
-			},
-			connectTimeout: timeout,
-			requestTimeout: timeout,
-		});
-	}
-
-	return await sql.connect(connectionInfo.connectionString);
-};
-
-const isEmail = name => /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(name || '');
-
-const getHostName = url => (url || '').split('.')[0];
+const { getConnectionClient } = require('./helpers/connection');
 
 const PERMISSION_DENIED_CODE = 297;
 
@@ -1024,95 +887,6 @@ const mapResponse = async (response = Promise.resolve({})) => {
 	const resp = await response;
 	return resp.recordset ? resp.recordset : resp;
 };
-
-const getTokenByMSAL = async ({ connectionInfo, redirectUri, clientId, tenantId, logger }) => {
-	try {
-		const pca = new msal.PublicClientApplication(getAuthConfig({ clientId, tenantId, logger }));
-		const tokenRequest = {
-			code: connectionInfo?.externalBrowserQuery?.code || '',
-			scopes: ['https://database.windows.net//.default'],
-			redirectUri,
-			codeVerifier: connectionInfo?.proofKey,
-			clientInfo: connectionInfo?.externalBrowserQuery?.client_info || '',
-		};
-
-		const responseData = await pca.acquireTokenByCode(tokenRequest);
-
-		return responseData.accessToken;
-	} catch (error) {
-		logger.log('error', { message: error.message, stack: error.stack, error }, 'MFA MSAL auth error');
-		return '';
-	}
-};
-
-const getAgent = ({ reject, cert, key } = {}) => {
-	return new https.Agent({ cert, key, rejectUnauthorized: !!reject });
-};
-
-const getTokenByAxios = async ({ connectionInfo, tenantId, redirectUri, clientId, logger, agent }) => {
-	try {
-		const params = new URLSearchParams();
-		params.append('code', connectionInfo?.externalBrowserQuery?.code || '');
-		params.append('client_id', clientId);
-		params.append('redirect_uri', redirectUri);
-		params.append('grant_type', 'authorization_code');
-		params.append('code_verifier', connectionInfo?.proofKey);
-		params.append('resource', 'https://database.windows.net/');
-
-		const responseData = await axios.post(`https://login.microsoftonline.com/${tenantId}/oauth2/token`, params, {
-			headers: {
-				'Content-Type': 'application/x-www-form-urlencoded',
-			},
-			...(agent && { httpsAgent: agent }),
-		});
-
-		return responseData?.data?.access_token || '';
-	} catch (error) {
-		logger.log('error', { message: error.message, stack: error.stack, error }, 'MFA Axios auth error');
-		return '';
-	}
-};
-
-const getTokenByAxiosExtended = ({ ...params }) => getTokenByAxios({ ...params, agent: getAgent() });
-
-const getToken = async ({ connectionInfo, tenantId, clientId, redirectUri, logger }) => {
-	const axiosExtendedToken = await getTokenByAxiosExtended({
-		connectionInfo,
-		clientId,
-		redirectUri,
-		tenantId,
-		logger,
-	});
-	if (axiosExtendedToken) {
-		return axiosExtendedToken;
-	}
-
-	const msalToken = await getTokenByMSAL({ connectionInfo, clientId, redirectUri, tenantId, logger });
-	if (msalToken) {
-		return msalToken;
-	}
-
-	const axiosToken = await getTokenByAxios({ connectionInfo, clientId, redirectUri, tenantId, logger });
-	if (axiosToken) {
-		return axiosToken;
-	}
-};
-
-const getAuthConfig = ({ clientId, tenantId, logger }) => ({
-	system: {
-		loggerOptions: {
-			loggerCallback(loglevel, message) {
-				logger.log(message);
-			},
-			piiLoggingEnabled: false,
-			logLevel: msal.LogLevel.Verbose,
-		},
-	},
-	auth: {
-		clientId,
-		authority: `https://login.microsoftonline.com/${tenantId}`,
-	},
-});
 
 const getDescriptionComments = async ({ client, dbName, schema, entity, logger }) => {
 	const currentDbConnectionClient = await getClient({
